% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/multiply.R
\name{divide}
\alias{divide}
\alias{multiply}
\title{Multiplies two numbers}
\usage{
divide(x, y)

multiply(x, y)
}
\arguments{
\item{x, y}{Scalar values to multiply or divide}
}
\description{
Performs multiplication and division of two integer or numeric scalars.
}
\examples{
multiply(2, 3)

}
\seealso{
\code{\link[=prod]{prod()}}
}
