% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic.drat-package.r
\docType{package}
\name{tic.drat-package}
\alias{tic.drat}
\alias{tic.drat-package}
\title{tic.drat: Example Package for 'tic'}
\description{
Building a package with pkgdown documentation.
}
\details{
For demonstration of the \pkg{tic} package.
}
\examples{
# Show timestamp
Sys.time()
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/tic.drat}
  \item \url{http://krlmlr.github.io/tic.drat}
  \item Report bugs at \url{https://github.com/krlmlr/tic.drat/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
